requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
pydantic==2.9.2
scikit-learn==1.5.2
pandas==2.2.3
numpy==1.26.4
joblib==1.4.2
python-multipart==0.0.12
openai
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
httpx
